全てのメッセージを取得するメソッドを切り出し
