+++ fuel/app/classes/model/user.php
@@ -41,16 +41,6 @@ class model_user extends \Orm\Model
 
     protected static $_table_name = 'users';
 
-    static public function message_find_by_userId_desc_by_createdAt($user_id)
-    {
-        return Model_Message::find('all', array(
-            'where' => array(
-                array('user_id', $user_id),
-            ),
-            'order_by' => array('created_at' => 'desc'),
-        ));
-    }
-
     static public function isNotYetExistedScreenName($check)
     {
         $result = self::find('all', array(
